fix(api): was calling upload through sftp twice. Big yikes if I didn't see this
Part of ENG-850
